refactor admin tutorials page with hooks
