+++ frontend/src/components/dashboard/admin/tutorials/RejectTutorialModal.js
@@ -1,0 +1,12 @@
+import RejectionReasonModal from "@/components/common/RejectionReasonModal";
+
+export default function RejectTutorialModal({ isOpen, onClose, onConfirm, t }) {
+  return (
+    <RejectionReasonModal
+      isOpen={isOpen}
+      onClose={onClose}
+      onConfirm={onConfirm}
+      title={t("reject_title")}
+    />
+  );
+}
